[ui] Split AutomaterializeMiddlePanel
