+++ js_modules/dagit/packages/core/src/assets/AutoMaterializePolicyPage/__stories__/AutomaterializeMiddlePanelNoPartitions.stories.tsx
@@ -2,33 +2,17 @@ import {MockedProvider} from '@apollo/client/testing';
 import * as React from 'react';
 
 import {RunStatus} from '../../../graphql/types';
-import {AutomaterializeMiddlePanel} from '../AutomaterializeMiddlePanel';
+import {AutomaterializeMiddlePanelNoPartitions} from '../AutomaterializeMiddlePanelNoPartitions';
 import {
   Evaluations,
   SINGLE_MATERIALIZE_RECORD_NO_PARTITIONS,
-  SINGLE_MATERIALIZE_RECORD_WITH_PARTITIONS,
 } from '../__fixtures__/AutoMaterializePolicyPage.fixtures';
 import {buildRunStatusOnlyQuery} from '../__fixtures__/RunStatusOnlyQuery.fixtures';
 
 // eslint-disable-next-line import/no-default-export
-export default {component: AutomaterializeMiddlePanel};
+export default {component: AutomaterializeMiddlePanelNoPartitions};
 
 const path = ['test'];
-
-export const WithPartitions = () => {
-  return (
-    <MockedProvider mocks={[Evaluations.Single(path)]}>
-      <div style={{width: '800px'}}>
-        <AutomaterializeMiddlePanel
-          assetKey={{path}}
-          assetHasDefinedPartitions
-          maxMaterializationsPerMinute={1}
-          selectedEvaluation={SINGLE_MATERIALIZE_RECORD_WITH_PARTITIONS}
-        />
-      </div>
-    </MockedProvider>
-  );
-};
 
 export const WithoutPartitions = () => {
   return (
@@ -36,9 +20,7 @@ export const WithoutPartitions = () => {
       mocks={[Evaluations.Single(path), buildRunStatusOnlyQuery('abcdef12', RunStatus.STARTED)]}
     >
       <div style={{width: '800px'}}>
-        <AutomaterializeMiddlePanel
-          assetKey={{path}}
-          assetHasDefinedPartitions={false}
+        <AutomaterializeMiddlePanelNoPartitions
           maxMaterializationsPerMinute={1}
           selectedEvaluation={SINGLE_MATERIALIZE_RECORD_NO_PARTITIONS}
         />
